moved cmd main to cmd client
